requirements.txt
torch
sentence-transformers
faiss-cpu
transformers
scikit-learn
fastapi
uvicorn
tabulate
sentencepiece
